Add new confirm library in specialties.jsx

+++ src/components/admin_panel/Specialties.jsx
@@ -1,5 +1,6 @@
 import { useEffect, useState } from "react"
 import { useLanguage } from "../context/LanguageContext"
+import Swal from "sweetalert2";
 export const Specialties = () => {
     const [specialties, setSpecialties] = useState(null)
     const token = localStorage.getItem("token");
@@ -9,7 +10,19 @@ export const Specialties = () => {
             .then(res => res.json())
             .then(data => setSpecialties(data.specialties))
     }, [])
-    const handleDeleteSpecialty = (id) => {
+    const handleDeleteSpecialty = async (id) => {
+        const result = await Swal.fire({
+                    title: 'Eliminar especialidad?',
+                    text: 'Esta acción no se puede deshacer',
+                    icon: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#3085d6',
+                    cancelButtonColor: '#d33',
+                    confirmButtonText: 'Sí, eliminar',
+                    cancelButtonText: 'Cancelar'
+        });
+        
+        if (!result.isConfirmed) return;
         fetch("http://localhost:3000/dashboard/specialty/", {
             method: "DELETE",
             headers: {
@@ -23,6 +36,18 @@ export const Specialties = () => {
 
     const createSpecialty = async (e) => {
         e.preventDefault();
+        const result = await Swal.fire({
+                    title: 'Crear especialidad?',
+                    text: 'Esta acción no se puede deshacer',
+                    icon: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#3085d6',
+                    cancelButtonColor: '#d33',
+                    confirmButtonText: 'Sí, crear',
+                    cancelButtonText: 'Cancelar'
+        });
+        
+        if (!result.isConfirmed) return;
         const specialty = e.target.specialty.value;
         const description = e.target.desc.value;
         const res = await fetch("http://localhost:3000/dashboard/specialty/", {
